DOCSP-27882 Atlas search page (#81)

+++ source/includes/code-examples/Address.cs
@@ -3,7 +3,7 @@ public class Address
     public string Building { get; set; }
 
     [BsonElement("coord")]
-    public float[] Coordinates { get; set; }
+    public double[] Coordinates { get; set; }
 
     public string Street { get; set; }
